[REFACTOR] Certik Audit Recommendation - Exhibit 19
Description:
Within the function “rewardPoolBalanceIncreased”, which is callable by any reward pool
provider, an unsafe addition is conducted on the “tokensToClaim” of a “snapshot” based on user
input.

Recommendations:
At its current state, it is possible to nullify the rewards of a snapshot by providing a “uint128
amount” that would cause the “tokensToClaim” member of the “DividendsSnapshot” struct to
overflow. We advise that the addition is checked against a potential overflow either manually or
by utilizing a safe library like “SafeMath” from OpenZeppelin.