[CHANGED] Certik Audit - Exhibit 7 (round 3)
[MINOR] Description:
The “start” function of the contract makes sure that the owner of the contract transmits the necessary funds to conduct the reward payouts and sets the “startTimestamp” variable to “now”.

Recommendations:
There is no guard against calling the “start” function twice so it is possible to “reset” the contract’s “startTimestamp” at any time. As “startTimestamp” is used to get the current cycles in numerous places, if the owner maliciously spams the contract with a “start” transaction and “withdrawRewardsPool” transaction to recoup the amount sent on each “start” transaction it is possible to freeze the contract in numerous sensitive functions, such as “unstakeNft”, as well as affect the logic of the contract in general.

Contracts are meant to give access to their owners on a “need-to” basis rather than an administrative paradigm. This function should be guarded against its execution occurring once by checking whether the “startTimestamp” variable is zero.