[CHANGED] Certik Audit - Exhibit 13 (round 3)
[MINOR] Description:
The lines of code above pertain to a logical misconception that multiplying by a big enough number, conducting division afterwards with the value we desire and then dividing by the same big enough number would yield higher numerical precision than doing the division directly.

Recommendations:
Solidity rounds numbers down and as such, this type of multiplication and subsequent division is ineffectual. As a side-effect, this also means that the full reward for a cycle will also never be properly divided unless the remainder is kept and handled on a case-by-case basis.

We propose the removal of the superfluous multiplications / divisions and the reconsiderance of how the on-chain dividend mechanism is handled as they are notoriously difficult to implement in an optimized and correct fashion. In the context of rewards, since the administrator is able to withdraw the reward token there is no apparent harm in leaving the remainder in-balance.